wordle-clone/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
firebase-admin==6.7.0
python-dotenv==1.0.1 